chore: owner 필드 변경 및 유효 접근 확인 로직 구현 - Store: User owner를 UUID owner_id로 변경 - StoreService: checkValidAccess() 메서드 추가

+++ src/main/java/com/team6/backend/store/application/service/StoreService.java
@@ -9,6 +9,7 @@ import com.team6.backend.store.domain.entity.Store;
 import com.team6.backend.store.domain.repository.StoreRepository;
 import com.team6.backend.store.presentation.dto.request.StoreRequest;
 import com.team6.backend.store.presentation.dto.response.StoreResponse;
+import com.team6.backend.user.domain.entity.Role;
 import com.team6.backend.user.domain.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -31,10 +32,10 @@ public class StoreService {
     @Transactional
     public StoreResponse createStore(StoreRequest request) {
         UUID userId = securityUtils.getCurrentUserId();
-        User owner = userRepository.findById(userId)
+        userRepository.findById(userId)
                 .orElseThrow(() -> new ApplicationException(CommonErrorCode.RESOURCE_NOT_FOUND));
 
-        Store store = new Store(owner, request.getCategoryId(), request.getAreaId(), request.getName(), request.getAddress());
+        Store store = new Store(userId, request.getCategoryId(), request.getAreaId(), request.getName(), request.getAddress());
         Store saved =  storeRepository.save(store);
         return new StoreResponse(saved);
     }
@@ -59,6 +60,7 @@ public class StoreService {
     @Transactional
     public StoreResponse updateStore(UUID storeId, StoreRequest request) {
         Store store = findStoreById(storeId);
+        checkValidAccess(store);
         store.update(request.getCategoryId(), request.getAreaId(), request.getName(), request.getAddress());
         return new StoreResponse(store);
     }
@@ -66,12 +68,14 @@ public class StoreService {
     @Transactional
     public void deleteStore(UUID storeId) {
         Store store = findStoreById(storeId);
+        checkValidAccess(store);
         storeRepository.delete(store);
     }
 
     @Transactional
     public StoreResponse hideStore(UUID storeId) {
         Store store = findStoreById(storeId);
+        checkValidAccess(store);
         store.hideStore();
         return new StoreResponse(store);
     }
@@ -79,5 +83,16 @@ public class StoreService {
     private Store findStoreById(UUID storeId) {
         return storeRepository.findById(storeId)
                 .orElseThrow(() -> new ApplicationException(StoreErrorCode.STORE_NOT_FOUND));
+    }
+
+    /* 사용자의 Role이 Customer이거나, 자신이 Owner가 아닌 Store에 접근할 때 예외를 발생시킵니다. */
+    private void checkValidAccess(Store store) {
+        Role role = securityUtils.getCurrentUserRole();
+        if (role == Role.CUSTOMER)
+            throw (new ApplicationException(CommonErrorCode.FORBIDDEN));
+        else if (role == Role.OWNER) {
+            if (store.getOwnerId() != securityUtils.getCurrentUserId())
+                throw (new ApplicationException(CommonErrorCode.FORBIDDEN));
+        }
     }
 }
